+++ src/ring/Ring.hpp
@@ -24,7 +24,7 @@ namespace coContext {
 
         auto registerSelfFileDescriptor(std::source_location sourceLocation = std::source_location::current()) -> void;
 
-        auto registerCpuAffinity(std::uint64_t cpuSetSize, const cpu_set_t *cpuSet,
+        auto registerCpuAffinity(std::size_t cpuSetSize, const cpu_set_t *cpuSet,
                                  std::source_location sourceLocation = std::source_location::current()) -> void;
 
         auto registerSparseFileDescriptor(std::uint32_t count,
